Abstract of this project
We are developing a Tech Services management system. In user page, users can make a request of technical services. And operators will attain the technician who is not have an unfinished job. In admin pages admin can add new categories, operators, technician etc. We used almost all of the facilities of Identity options such as mailing, reactivation codes, refreshing user information etc. By the time this project will be finished we are going to use google maps to choose location for user, and also users will be able to share a photo of their problem etc.